# swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "SystemGatewayAPI",
    "version": "1.0"
  },
  "paths": {
...
    "/api/Therapist/Register": {
      "post": {
        "tags": [
          "Therapist"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Therapist"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Therapist"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Therapist"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Therapist/{token}": {
      "put": {
        "tags": [
          "Therapist"
        ],
        "parameters": [
          {
            "name": "Token",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Therapist"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Therapist"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Therapist"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "Therapist"
        ],
        "parameters": [
          {
            "name": "Token",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "get": {
        "tags": [
          "Therapist"
        ],
        "parameters": [
          {
            "name": "Token",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Therapist/{token}/Reject/{patientEmail}": {
      "get": {
        "tags": [
          "Therapist"
        ],
        "parameters": [
          {
            "name": "token",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "patientEmail",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Therapist/{token}/Accept/{patientEmail}": {
      "get": {
        "tags": [
          "Therapist"
        ],
        "parameters": [
          {
            "name": "token",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "patientEmail",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
    "/api/Therapist/{token}/Patients": {
      "get": {
        "tags": [
          "Therapist"
        ],
        "parameters": [
          {
            "name": "Token",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Therapist/{token}/Patients/{email}": {
      "get": {
        "tags": [
          "Therapist"
        ],
        "parameters": [
          {
            "name": "Token",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "patientEmail",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "email",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Therapist/{token}/Patients/{email}/Modules": {
      "get": {
        "tags": [
          "Therapist"
        ],
        "parameters": [
          {
            "name": "Token",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "patientEmail",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "email",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Therapist/{token}/Patients/{email}/Modules/{ModuleId}": {
      "get": {
        "tags": [
          "Therapist"
        ],
        "parameters": [
          {
            "name": "Token",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "patientEmail",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "ModuleId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "email",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Therapist/{Token}/Patients/{PatientEmail}/Modules/{ModuleId}": {
      "put": {
        "tags": [
          "Therapist"
        ],
        "parameters": [
          {
            "name": "Token",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "PatientEmail",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "ModuleId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Module"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Module"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Module"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "get": {
        "tags": [
          "Therapist"
        ],
        "parameters": [
          {
            "name": "Token",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "PatientEmail",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "ModuleId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Therapist/{Token}/Patients/{PatientEmail}/Modules/{ModuleId}/Version/{VersionId}": {
      "put": {
        "tags": [
          "Therapist"
        ],
        "parameters": [
          {
            "name": "Token",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "PatientEmail",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "ModuleId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "VersionId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "DataPoint": {
        "type": "object",
        "properties": {
          "sectionName": {
            "type": "string",
            "nullable": true
          },
          "isDataEditable": {
            "type": "boolean"
          },
          "content": {
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "Module": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "moduleData": {
            "$ref": "#/components/schemas/ModuleVersion"
          }
        },
        "additionalProperties": false
      },
...
      "ModuleVersion": {
        "type": "object",
        "properties": {
          "versionId": {
            "type": "string",
            "nullable": true
          },
          "applicationName": {
            "type": "string",
            "nullable": true
          },
          "dataStructure": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DataPoint"
            },
            "nullable": true
          },
          "htmlCard": {
            "type": "string",
            "nullable": true
          },
          "htmlDashboard": {
            "type": "string",
            "nullable": true
          },
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "checksum": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "Therapist": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "nullable": true
          },
          "password": {
            "type": "string",
            "nullable": true
          },
          "firstName": {
            "type": "string",
            "nullable": true
          },
          "lastName": {
            "type": "string",
            "nullable": true
          },
          "phoneNumber": {
            "type": "string",
            "nullable": true
          },
          "address": {
            "type": "string",
            "nullable": true
          },
          "age": {
            "type": "integer",
            "format": "int32"
          },
          "profilePicture": {
            "type": "string",
            "nullable": true
          },
          "credentials": {
            "type": "string",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "acceptedPatients": {
            "uniqueItems": true,
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          },
          "requestedPatients": {
            "uniqueItems": true,
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
    }
  }
}